Bacon

+++ src/runner.js
@@ -13,6 +13,7 @@ import colors from 'colors'
 import path from 'path'
 import minimist from 'minimist'
 import ms from 'ms'
+import fs from 'fs'
 
 // node dist/runner.js --num-containers 2 --num-peers 10 --signaling-url ws://192.168.1.144:8080/socketserver
 
@@ -32,7 +33,7 @@ const SIGNALING_URL = args['signaling-url'] || 'ws://178.62.51.86:8080/socketser
 const TIMEOUT = args['timeout'] || ms('5m')
 const MODE = args['mode'] || 'full' // mode can be either 'full', 'spawn' or 'walker'
 const DOCKER_IMAGE_ID = `webrtc/${uuid.v1()}`
-const spawns = []
+const DATA_SHEET_PATH = path.join(__dirname, '..', 'data', args['output-file'] || `${Date.now()}_test_result.data`)
 
 if (MODE === 'full') {
   async.series([
@@ -73,14 +74,12 @@ if (MODE === 'walker') {
 
 function createDockerImage (cb) {
   const child = spawn('docker', ['build', '-t', DOCKER_IMAGE_ID, '.'])
-  spawns.push(child)
   child.on('exit', cb)
 }
 
 function createBootPeer (cb) {
   const UUID = uuid.v1()
-  const child = spawn('docker', ['run', '--rm', DOCKER_IMAGE_ID, 'test', 'peer', SIGNALING_URL, 1, UUID])
-  spawns.push(child)
+  spawn('docker', ['run', '--rm', DOCKER_IMAGE_ID, 'test', 'peer', SIGNALING_URL, 1, UUID])
   setTimeout(cb, 5000)
 }
 
@@ -94,7 +93,6 @@ function runContainer (currentNum, type, cb) {
 
   // Spawn child process
   const child = spawn('docker', ['run', '--rm', DOCKER_IMAGE_ID, 'test', type, SIGNALING_URL, NUM_PEERS, UUID])
-  spawns.push(child)
   child.stdout.on('data', function (data) {
     console.log(data.toString())
     if (data.toString().indexOf('**NEXT**') !== -1) {
@@ -105,8 +103,8 @@ function runContainer (currentNum, type, cb) {
 
 function startWalker (cb) {
   const child = spawn('docker', ['run', '--rm', DOCKER_IMAGE_ID, 'test', 'walker', SIGNALING_URL])
-  spawns.push(child)
   let numConnections = 0
+  let durations = []
   let timeTotal = 0
   let timeMin = 0
   let timeMax = 0
@@ -135,6 +133,10 @@ function startWalker (cb) {
           if (duration < timeMin) timeMin = duration
           if (duration > timeMax) timeMax = duration
 
+          durations.push(duration)
+
+          fs.appendFile(DATA_SHEET_PATH, `${duration}\n`, (err) => {})
+
           timeTotal += duration
 
           numConnections++
@@ -144,23 +146,29 @@ function startWalker (cb) {
         }
 
         prevTime = timestamp
+
+        if (durations.length === NUM_PEERS * NUM_CONTAINERS) {
+          const mean = calculateMean(timeTotal, numConnections)
+          const variance = calculateVariance(durations, mean)
+          const standardDeviation = calculateStandardDeviation(variance)
+
+          console.log('')
+          console.log('-------- ⚡️  Test completed ⚡️ --------')
+          console.log('')
+          console.log(`Number of connection handovers: ${colors.yellow.bold(numConnections)}`)
+          console.log(`Min (fastest handover):         ${colors.yellow.bold(timeMin.toFixed(2) + ' ms')}`)
+          console.log(`Max (slowest handover):         ${colors.yellow.bold(timeMax.toFixed(2) + ' ms')}`)
+          console.log(`Mean:                           ${colors.green.bold.underline(`${mean.toFixed(2)}`)}`)
+          console.log(`Variance:                       ${colors.green.bold.underline(`${variance.toFixed(2)}`)}`)
+          console.log(`Standard deviation:             ${colors.green.bold.underline(`${standardDeviation.toFixed(2)}`)}`)
+          console.log('')
+
+          child.kill()
+          cb()
+        }
       })
     }
   })
-
-  setTimeout(() => {
-    console.log('')
-    console.log('-------- ⚡️  Test completed ⚡️ --------')
-    console.log('')
-    console.log(`Number of connection handovers: ${colors.yellow.bold(numConnections)}`)
-    console.log(`Fastest connection setup time:  ${colors.yellow.bold(timeMin.toFixed(2) + ' ms')}`)
-    console.log(`Slowest connection setup time:  ${colors.yellow.bold(timeMax.toFixed(2) + ' ms')}`)
-    console.log(`Avg. connection setup time:     ${colors.green.bold.underline(((timeTotal / (numConnections)).toFixed(2)) + ' ms')}`)
-    console.log('')
-
-    child.kill()
-    cb()
-  }, TIMEOUT)
 }
 
 function clean () {
@@ -180,4 +188,21 @@ function clean () {
 
 function sleep (millis, cb) {
   setTimeout(cb, millis)
+}
+
+function calculateMean (total, number) {
+  return (total / number)
+}
+
+function calculateVariance (inputs, mean) {
+  const number = inputs.length
+  let total = 0
+  for (const input of inputs) {
+    total += Math.pow((input - mean), 2)
+  }
+  return (total / number)
+}
+
+function calculateStandardDeviation (variance) {
+  return Math.sqrt(variance)
 }
